requirements.txt
Django>=3.1.1,<3.2.0
djangorestframework>=3.11.1,<3.12.0
flake8>=3.8.3,<3.9.0
psycopg2>=2.8.5,<2.9.0
